Correctly use name of street relations
If there is a osm_relation with the type 'street' or 'associatedStreet',
the name of the relation was applied to the members of the relation.
Since only 'name:en' was imported for relation names, it lead to many
cases where the name of the relation was NULL, which was applied to the
members. Like this, the name of certain streets was overwritten with
NULL and later deleted.

I applied two fixes for this:

1. names of relations are only taken into account when it is not empty
2. imposm now imports all name tags for relations and these are later
   used to determine the name of the relation